[herd] Explicit address ports, AArch64 semantics
We replace explicit data ports by explicit address ports.

Namely only one of those input ports need to be computed as the other
can be deduced from it (as DATA = Rreg \ ADDR for instance).

More specifically, the set  ADDR is pre-defined
for the Cat interpreter. The standard library file
`stdlib.cat` define `NDATA` as an alias for `ADDR`
and computes `DATA`, for the sake of (potential)
backward compatibility.

Notice that the definition of `DATA` for AArch64
(in `aarch64deps.cat`) is more specific and overrides
the naive default.