[herd] Rework input ports
We replace explicit data ports by
explicit data and address ports.

As a consequence semantics now take
three kind of input ports into account:
address, data, and none of those.

To maintain backard compatibility
`stdlib.cat` define `NDATA` as an alias for `ADDR`

All semantics are modifiled accordingly,
although this is meaningfull only for
models for which address and data dependencies
are significant.