[herd] Fix aarch64 <-> ASL interface
Architecture flags NZCV have different incarnations:
  + For ASL, global variable `_NZCV`, holding a bitvector.
  + For AArch64, register holding an integer.

This commit implements the cast from integer to bitvector
that was absent from the AArch64 to ASL interface.

Notice that the translation from ASL to AArch64 is performed differently:
writes to `_NZVC` are intercepted during ASL interpretation,
so as to generate the appropriate write effect, with integer value.