[dot2desc] Add support for memory accesses with no specified address registers
This is needed in the case of SVE instructions, where the addressing is not done
through one register, but is rather a combination of registers and immediates.
The tool does not attempt to describe that kind of addressing at the moment.
Instead, it decides to omit it and only mention the memory location