[herd] Fix some problem of the toplogical solver.
  1. Iterate solver step on dependency cycles.
  2. Add an ultimate substitution step.

For 1. the solver step has to be iterated because one such step
can yeild a new solution that would apply to some constraint
that appears before it in the cycle represented as a list.
Typically: V1:=F(V2); V2:=F(V1) when V1 <- cst1 is part of the
prior solution will require two solving steps:
  a) cst1:=F(V2) + new solution V2 <- cst2
  b) check cst1:=F(cst2)

For 2. In some rare occasions, a variable S can have two "definitions",
_i.e._ there are two equations `S:=...`. The definitions are
independent, so that the topological sort can handle them in any
order, If the second "definition" is solved, yielding a substitution
`S <- Value`, this substitution is not performed on the previous
"definition" and the second definition seen as a constraint is
lost. The fix consists in applying the complete substitution to all
remaining equations once, at the very end of the topological solver.

Co-authored-by: Hadrien Renaud <hadrien.renaud2@arm.com>