[herd] Complete tracking of the NZCV "register"
Instruction that write into the NZCV "register" must
record this as follows:
  1. Include the `NZCV` into the kill list of the instruction,
    (function `killed` in AArch64Arch_herd.ml`)
  2. Include the new binding as a result of instruction
     interpretation.

Notice that 1. or 2. alone are enough, 1. is simpler; while 2: leads
to potential optimisations of for interpreting the instructions
that follow. Having both enhances code robustness, especially in
the presence of alternative means of instruction interpretation
such as ASL.

Also notice a quick way to implement 1 is to kill all registers as
we do for all morello instructions.

As far as we have checked instructions that write the NZCV flags are:
Some operations, namely ADDS, SUBS, ANDS, BICS,
the SVE WHILE.. instructions, the Morello CSEAL instruction.
and the MSR NZCV,... instruction.