[tools] Fix description of the set-lifting relation by **miaou**.
In the general case, the relation `[evts]` description should refer to
two events. It can be done by replacing `[evts]` by `[evts] & id` that
will be described as "E1 in evts; E1 and E2 are the same events.
Nevertheless, by careful selection of event numbers this can be
avoided in sequences. For instance  `a;[evts];b` can be described as
"E1 a-relates E3; E3 is in evts; E3 b-relates E2`.