[asl+herd] Implement a exclusive monitor that satisfies comments
In VMSA mode both the load reserve (LDXR) and store conditional (STXR)
reads the page table entry for the effective address twice. Some
care is taken to compare the resulting physical addesses
and to reject candidate execution when they differ, by using
our `CheckProp` primitive.

Monitor implementation very much differ in ASL and  VMSA+ASL modes.
Hence, they are located in the physmem-std.asl and physmem-vmsa.asl
files. Namely, the ASL implementation of exclusives monitors
operate on virtual addresses for non-VMSA and on physical
addresses for VMSA. Then, it looks natural to have those
functions in physmem-std.asl for non-VMSA and in
physmem-vmsa.asl for VMSA.