[herd,asl] Store exclusive cancels reservation
This behaviour is the one of the AArch64 implementation.
However the manual suggests that this cancelation is not mandatory:

+ The store is guaranteed to succeed only if the PA accessed
is marked as exclusive access for the requesting PE and both
the local monitor and the global monitor state machines for
the requesting PE are in the Exclusive Access state. In this case:
  ...
  - The final state of the global monitor state machine for the
    requesting PE is IMPLEMENTATION DEFINED.
  ...