[herd,test] Test diymicro generated tests in asl mode
The new test `make diymicro-test-aarch64-asl` is
not part of the  default test suite `make test`.
However, it is part of  the complete test suite
`make test-all`.

Notice that two tests are not run purposely, because their asl
output differ from the standard one as regards the
number of executions.