[litmus] Change handling of the self variant
This variant impacts (1) the C code of a test and (2) gcc compilation flags.
Compilation flags are now changed
as soon as one test is compiled in self mode.
This departs from previous behaviour where compilation
flags were changed only when the variant was specified on
the command line.

Change is minor and only applies when
 1. `-variant self` is *not* specified on the command line
 2. Some of the compiled tests meta-data specifies `Variant=self`.