[litmus] Refactor flag collection into dedicated helper
Flags are currently collected as part of the run_tests
function. However, we will soon need access to flags earlier in the
process, specifically when dumping run commands.

To support this, let’s extract the flag collection logic into a
dedicated helper function. Since we are not interested in anything
other than flags drop any output and do not care about hashes.

Signed-off-by: Vladimir Murzin <vladimir.murzin@arm.com>