[gen] Remove `fullkvm` and combined annotation in kvm.
Since the difference between `fullkvm` and `kvm` is the allowed syntax
in annotation, for example `PteAF.DB.DBM` is only allowed the former.
However the same can be achieved by `PteAF PteDB PteDBM` in both mode,
we remove fullkvm and also remove the combined syntax such as
`PteAF.DB`. This change reduces the booting time of `diy` as there is no
need to build a syntax look-up table for the combined annotation anymore.