[herd,asl,vmsa] New implementation of the failing CAS.
In pure VMSA mode, there are are multiple executions of
the CAS instruction. We override one function `CreateAccDescAtomicOp`
two spilt the current execution in two:

  1. An execution with the same access decriptor as before
     and that will thus behave as before.

  2. An execution with descriptor field `write` being `FALSE` that
     will behave as a load.

This, with an additional check *not* to write when
the access descriptor `write` field is `FALSE` in `PhysMemWrite`
apparently suffice.

Interestingly the new execution is completly discarded when the CAS
comparison succeeds and only half discarded when the CAS comparison
fails.