BPF: Add litmus tests illustrating "Atomicity of RMWs"
The principle of "Atomicity of RMWs" is enforced via the rule "Atomic"
for RMWs which generate two memory events, while it is enforced via the
rule "Coherence" for RMWs which generate a single memory event.

Add litmus tests "successful-cmpxchg-atomicity" and "xchg-atomicity" to
illustrate such rules and principle (test "LockTwice" also illustrates
"Coherence" and "Atomicity of RMWs").

Signed-off-by: Andrea Parri <parri.andrea@gmail.com>
Suggested-by: Puranjay Mohan <puranjay@kernel.org>