[AARCH-23849,herd] Replace rf-reg with lrrs in aarch64deps.cat
 This change implements AARCH-23849 as published in the Arm
 Architecture Reference Manual for A-profile architecture: Known
 issues in Issue L.b.

This change replaces the relation of reads-from-register (rf-reg) with
a new relation local register read successor (lrrs). Local register
read successor builds on top of loc that applies between Register
Effects to the same Register and po that applies between Effects that
appear in program order.

Local register read successor can be derived statically from the
litmus test and without needing herd7 to compute the the reads from
relation.

Reads-from-register is used in dependencies and replacing it with
local register read successor is not meant to change the behaviour of
any tests that rely on dependecies that involve general purpose
registers only.

Signed-off-by: Nikos Nikoleris <nikos.nikoleris@arm.com>