[internal] Add option `-checkobs` to the regression test application.
This option selects a mode that reduces the validity of a test
to the quality of the text component of the `Observation` field
in logs (_i.e._ Always/Sometimes/Never).

The intent is to abstract away other log components, such as
the number of executions, which are less significant. We thought
this abstraction to be necessary to check that running the tests
in `herd/tests/instructions/AArch64` in ASL mode yield the same
results as in non-ASL mode. In effect, (see next commit)
this was not useful :  runnable test logs  match perfectly.