[internal] Add options to the regression test application.
  + `-verbose` will reassure us by printing some diagnostics whenever
  a test yields some output.

  + `-checkobs` This option selects a mode that reduces the validity
  of a test to the quality of the text component of the `Observation`
  field in logs (_i.e._ Always/Sometimes/Never).

  +`-checkstates` Reduce the validity check to final state equality.

The last to options apply to `herd_regression_test.exe` only The
intent is to abstract away other log components, such as the number of
executions, which are less significant. We thought this abstraction to
be necessary to check that running the tests in
`herd/tests/instructions/AArch64` in ASL mode yield the same results
as in non-ASL mode. In effect, (see next commit) this was not useful:
runnable test logs  match perfectly. Nevertheless we keep the changes
for future usage.