[herd] Delay check  that accesses are symbolic.
The simulator **herd7** should fail when a memory access
to a non-symbolic address is attempted. This pre-existing
(see PR #1513) check is now performed *after* the application
of the memory model.

As a result, tentative execution candidates that
perform such illegal accesses and that are
rejected by the model do *not* trigger failure
of the whole simulation, as expected.

We also add accesses to code through labels as symbolic,
a case overlooked by PR #1513.