[gen] Fix a problem on vmsa+moreedges, caused by `Shootdown` conflict.
When `-variant kvm` and `-moredeges true`, the booting stage of `diy*`
fails, due to a `pp` function, printing the same output/target
string despite different input data. Further it is because `Shootdown`
has `sync` and `nosync` mode, while in `nosync` mode, it still contains
domain information. This means, different `nosync` with different domain
will correspond to the same string by `pp` function. To solve the problem,
we introduce `ShootdownSync` and ShootDownNoSync`.

Also fix a problem `compare_fence` which previously did not exaust all
possibilty in particlar if both side is `CMO`. It now use the default
`stdlib.compare` which gives a order based on the data type defintiion.