[litmus7, aarch64] Replace NOP as start/stop marker in litmus assembly
`litmus7` (when using `-variant self`) needs to identify the boundaries
of the litmus assembly in the final binary. To achieve this, it wraps
the inline assembly with special markers and provides runtime
infrastructure to locate them.

Currently, the NOP instruction is used as a marker. Although NOPs can
also appear in the litmus assembly, `litmus7` handles this by counting
them so that they do not interfere with the markers.

However, we are about to emit an `.align` directive into the inline
assembly. This directive introduces NOPs as padding, which are outside
the `litmus7`’s visibility and can therefore interfere with the marker
detection.

To avoid this issue, switch the marker instruction from NOP to a less
ambiguous one, such as `MOV WZR, #0`.

Signed-off-by: Vladimir Murzin <vladimir.murzin@arm.com>