[litmus7, aarch64] Handle Instruction Abort
Instruction Abort comes with it's own Exception Code (EC), use that to
wire up with our fault_handler. Note, that we do not have any symbol
associated with the code, only lables can be used for that, thus do
not attempt to match fault address due to instruction abort with any
symbol.

Signed-off-by: Vladimir Murzin <vladimir.murzin@arm.com>