[litmus7,WIP] Initial support for translating addresses of instructions
Translating instruction addresses requires the litmus assembly to
reside at page-size granularity. We already control alignment within
the litmus assembly using the `.pagealign` directive, but we must also
handle alignment outside of it.

First, we need to ensure that the start of the litmus assembly is
page-aligned and that the first instruction of the litmus test is
placed at the beginning of the page. To achieve this, we implicitly
insert a `.pagealign` directive at the start of the litmus assembly -
between the start marker and the first instruction of the litmus
test. We also align the `prelude_size` to the page boundary to account
for the offset between the start marker and the first instruction.

Second, when a fault handler is present, we must ensure safe recovery
from faults. For this, we align everything following the litmus
assembly to page granularity by implicitly inserting a `.pagealign`
directive just before the stop marker.

Third, the compiler-emitted code is copied into a dedicated
page-aligned buffer for execution and modification. Copying the code
as-is could break page alignment guarantees since the emitted code is
not guaranteed to be page-aligned. Therefore, we enforce alignment of
the generated code.

Additionally, introduce logic to collect, save, manipulate, and
restore page table entries for labels present in the litmus test.

Signed-off-by: Vladimir Murzin <vladimir.murzin@arm.com>