[herd] Do not specifically check for non-symbolic addresses in VMSA mode
PR #1513 introduced a necessary but expensive check: loads
of which address is not a symbol lead to an extra execution
candidate that checks for non-symbolic addresses.

Such an extra-candidate is not needed in the VMSA case
because of the systematic application of primitives (such
as PTELoc or Offset) that fail on a non-symbolic argument.