[catalogue] Revamp the aarch64-ETS2 catalogue
Previously, the aarch64-ETS2 was a soft link to the aarch64-VMSA
catalogue. This meant that we would run the same tests as aarch64-VMSA
with the addition of -variant ets2. In addition, there was a separate
kinds.txt file because -variant ets2 meant that some tests would
change to being architecturally forbidden.

This change switches to a custom set of tests which are designed to
illustrate the most representative tests that become forbidden when
FEAT_ETS2 is implemented.

Signed-off-by: Nikos Nikoleris <nikos.nikoleris@arm.com>