[AARCH-24989,herd] Remove lfs from lob in aarch64hwreqs.cat
This change removes lfs from lob and as a consequence removes the now
unused definition of lfs itself. Then it simplifies the definition of
same-loc which previously had the added complexity of defining when an
MMU Fault Effect was to the same location as another Effect.

Removing lfs from lob changes the intended outcome of a number of
tests which are also updated and as as a result the corresponding
kinds file also changes.

Signed-off-by: Nikos Nikoleris <nikos.nikoleris@arm.com>