[AARCH-24987,herd] Relax bob in relation to Fault Effects
This change implements AARCH-24987 as published in the Arm
Architecture Reference Manual for A-profile architecture: Known issues
in Issue M.a.

This issue proposes multiple retrospective relaxations of the hardware
behavior in relation to the ordering requirements of instructions with
Release semantics.

In the Arm ARM issue M.a, section B2.3.7 "Ordering relations", the
definition of Barrier-ordered-before includes cases that mention Fault
Effects with Release semantics. These are Fault Effects generated by
instructions with Release semantics that raise a synchronous exception
due to a Data Abort. Arm has received feedback that the concept of Fault
Effects with Release semantics is not natural for hardware
implementations, and that in some implementations the release semantics
are not preserved when an instruction that specifies release semantics
raises a synchronous exception due to a Data Abort. For this reason, Arm
is proposing to relax the definition of Barrier-ordered-before and
remove all uses of Fault Effects with Release semantics. Note that:

- For hardware implementing FEAT_ETS3, this relaxation does not matter.
  The definition of ETS-ordered-before provides order in all cases of
  synchronous exceptions due to a Data Abort.

- For hardware implementing FEAT_ETS2, this relaxation only affects
  cases of instructions with release semantics that raise a data abort
  synchronous exception due to an MMU Permission fault.

In addition, the definition of Barrier-ordered-before includes two
clauses that reference Intrinsic Order Dependency which apply in the
case of Explicit Memory Effects generated by an LDIAPP or an STILP
instruction. These two clauses were inadvertently extended beyond the
original intent which is to specify the order of the Explicit Memory
Effects of LDIAPP and STILP to include ordering requirements in relation
to Implicit Tag Memory Effects. For this reason, Arm is proposing to
relax these two clauses.

Signed-off-by: Nikos Nikoleris <nikos.nikoleris@arm.com>